+++ minus_games_gui/src/minus_games_gui/messages/minus_games_gui_message.rs
@@ -13,7 +13,7 @@ pub(crate) enum MinusGamesGuiMessage {
     Init,
     // InitComplete(Arc<RwLock<tokio::sync::mpsc::Receiver<MinusGamesClientEvents>>>),
     InitWindow(()),
-    SetScale(f64),
+    SetScale(f32),
     InitComplete(()),
     Loading,
     Reload,
